Pull functions out of index and into modules
