+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
   entry: {
-    main: "./lib/index.js",
+    main: ["./lib/index.js"],
     test: "mocha!./test/index.js"
   },
   output: {
